+++ test/tests/functional/rs_mocks/connection_tests.js
@@ -666,3 +666,156 @@ exports['Successful connection to replicaset of 0 primary, 1 secondary and 1 arb
     }, 100)
   }
 }
+
+exports['Should print socketTimeout warning due to socketTimeout < haInterval'] = {
+  metadata: {
+    requires: {
+      generators: true,
+      topology: "single"
+    }
+  },
+
+  test: function(configuration, test) {
+    var ReplSet = configuration.require.ReplSet,
+      ObjectId = configuration.require.BSON.ObjectId,
+      Connection = require('../../../../lib/connection/connection'),
+      ReadPreference = configuration.require.ReadPreference,
+      Long = configuration.require.BSON.Long,
+      co = require('co'),
+      mockupdb = require('../../../mock');
+
+    // Contain mock server
+    var primaryServer = null;
+    var firstSecondaryServer = null;
+    var arbiterServer = null;
+    var running = true;
+    var electionIds = [new ObjectId(), new ObjectId()];
+
+    // Default message fields
+    var defaultFields = {
+      "setName": "rs", "setVersion": 1, "electionId": electionIds[0],
+      "maxBsonObjectSize" : 16777216, "maxMessageSizeBytes" : 48000000,
+      "maxWriteBatchSize" : 1000, "localTime" : new Date(), "maxWireVersion" : 4,
+      "minWireVersion" : 0, "ok" : 1, "hosts": ["localhost:32000", "localhost:32001", "localhost:32002"], "arbiters": ["localhost:32002"]
+    }
+
+    // Primary server states
+    var primary = [extend(defaultFields, {
+      "ismaster":true, "secondary":false, "me": "localhost:32000", "primary": "localhost:32000", "tags" : { "loc" : "ny" }
+    })];
+
+    // Primary server states
+    var firstSecondary = [extend(defaultFields, {
+      "ismaster":false, "secondary":true, "me": "localhost:32001", "primary": "localhost:32000", "tags" : { "loc" : "sf" }
+    })];
+
+    // Primary server states
+    var arbiter = [extend(defaultFields, {
+      "ismaster":false, "secondary":false, "arbiterOnly": true, "me": "localhost:32002", "primary": "localhost:32000"
+    })];
+
+    // Boot the mock
+    co(function*() {
+      primaryServer = yield mockupdb.createServer(32000, 'localhost');
+      firstSecondaryServer = yield mockupdb.createServer(32001, 'localhost');
+      arbiterServer = yield mockupdb.createServer(32002, 'localhost');
+
+      // Primary state machine
+      co(function*() {
+        while(running) {
+          var request = yield primaryServer.receive();
+          var doc = request.document;
+
+          if(doc.ismaster) {
+            request.reply(primary[0]);
+          }
+        }
+      }).catch(function(err) {
+        console.log(err.stack);
+      });
+
+      // First secondary state machine
+      co(function*() {
+        while(running) {
+          var request = yield firstSecondaryServer.receive();
+          var doc = request.document;
+
+          if(doc.ismaster) {
+            request.reply(firstSecondary[0]);
+          }
+        }
+      }).catch(function(err) {
+        console.log(err.stack);
+      });
+
+      // Second secondary state machine
+      co(function*() {
+        while(running) {
+          var request = yield arbiterServer.receive();
+          var doc = request.document;
+
+          if(doc.ismaster) {
+            request.reply(arbiter[0]);
+          }
+        }
+      }).catch(function(err) {
+        console.log(err.stack);
+      });
+    });
+
+    Connection.enableConnectionAccounting();
+    // Attempt to connect
+    var server = new ReplSet([
+      { host: 'localhost', port: 32000 },
+      { host: 'localhost', port: 32001 },
+      { host: 'localhost', port: 32002 }], {
+        setName: 'rs',
+        connectionTimeout: 3000,
+        socketTimeout: 2000,
+        haInterval: 5000,
+        size: 1
+    });
+
+    server.on('joined', function(_type) {
+      if(_type == 'arbiter' || _type == 'secondary' || _type == 'primary') {
+        // console.log("!!!!!!!!!!!!!!!!! joined :: " + _type)
+        // console.log("server.s.replicaSetState.secondaries = " + server.s.replicaSetState.secondaries.length)
+        // console.log("server.s.replicaSetState.arbiters = " + server.s.replicaSetState.arbiters.length)
+
+        if(server.s.replicaSetState.secondaries.length == 1
+          && server.s.replicaSetState.arbiters.length == 1
+          && server.s.replicaSetState.primary) {
+            test.equal(1, server.s.replicaSetState.secondaries.length);
+            test.equal('localhost:32001', server.s.replicaSetState.secondaries[0].name);
+
+            test.equal(1, server.s.replicaSetState.arbiters.length);
+            test.equal('localhost:32002', server.s.replicaSetState.arbiters[0].name);
+
+            test.ok(server.s.replicaSetState.primary != null);
+            test.equal('localhost:32000', server.s.replicaSetState.primary.name);
+
+            primaryServer.destroy();
+            firstSecondaryServer.destroy();
+            arbiterServer.destroy();
+            server.destroy();
+            running = false;
+
+            setTimeout(function() {
+              test.equal(0, Object.keys(Connection.connections()).length);
+              Connection.disableConnectionAccounting();
+              test.done();
+            }, 1000);
+          }
+      }
+    });
+
+    server.on('connect', function(e) {
+      server.__connected = true;
+    });
+
+    // Gives proxies a chance to boot up
+    setTimeout(function() {
+      server.connect();
+    }, 100)
+  }
+}